pyproject.toml
[tool.poetry]
name = "fastapiproject1"
version = "0.1.0"
description = ""
authors = ["Bazavod <eugeny.bazavod@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.44"
asyncpg = "^0.30.0"
databases = "^0.9.0"
python-multipart = "^0.0.20"
email-validator = "^2.3.0"
itsdangerous = "^2.2.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-jose = "^3.5.0"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.11"
pydantic = "^2.12.4"
argon2-cffi = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
